add code depending on whether showing shop or cart

+++ src/components/Cart.js
@@ -4,13 +4,29 @@ import "./Cart.css"
 import { useState,useEffect } from "react";
 
 const Cart=(props)=>{
-return (
-    <div className="cart">
-{props.cart.map(item=>{
-return <Item itemID={item.itemID} itemTitle={item.itemTitle} itemImage={item.itemImage} deleteItem={props.deleteItem} itemQuantity={item.itemQuantity}/>
-})}
-                </div>
-)
+    if (props.cartPage) {
+        return (
+            <div>
+            <Navbar/>
+            <div className="cart">
+           <h3># of items: {props.count }</h3>
+        {props.cart.map(item=>{
+        return <Item itemID={item.itemID} itemTitle={item.itemTitle} itemImage={item.itemImage} deleteItem={props.deleteItem} itemQuantity={item.itemQuantity}/>
+        })}
+                        </div>
+                </div>        
+        )
+    }
+else {
+    return (
+        <div className="cart">
+        <h3># of items: {props.count}</h3>
+    {props.cart.map(item=>{
+    return <Item itemID={item.itemID} itemTitle={item.itemTitle} itemImage={item.itemImage} deleteItem={props.deleteItem} itemQuantity={item.itemQuantity}/>
+    })}
+                    </div>
+    )
+}
 }
 
 export default Cart;
